moved MessageBoard component to VideoList.tsx
