+++ packages/frontend/pages/test.tsx
@@ -1,0 +1,14 @@
+import React from 'react'
+//import MessageBoard from '../components/MessageBoard'
+import VideoBoard from '../components/Video/VideoBoard'
+//import ImageBoard from '../components/ImageBoard'
+
+export default function test() {
+  return (
+    <>
+      <div>
+        <VideoBoard />
+      </div>
+    </>
+  )
+}
